pyproject.toml
[project]
name = "aiagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "groq>=0.11.0",
    "python-dotenv==1.1.0",
]
